fix(engine): resolve compilation errors in engine tests
- Remove duplicate downloadLoop function in download_engine.go
- Remove unused imports (encoding/hex, math/rand) from download_engine.go
- Fix torrent.InfoHash -> torrent.Info.InfoHash references
- Remove duplicate TestGetNextPiece from piece_manager_test.go
- Fix totalSize type (int -> int64) in engine test files

+++ internal/engine/download_engine_test.go
@@ -52,7 +52,7 @@ func TestNewDownloadEngine_CustomOutputDir(t *testing.T) {
 
 func TestGetNextPiece(t *testing.T) {
 	pieceLength := 16384
-	totalSize := 16384 * 5
+	totalSize := int64(16384 * 5)
 	numPieces := 5
 
 	pieceHashes := make([]byte, numPieces*20)
